+++ packages/components/manageable-list/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-manageable-list@0.12.0...@synerise/ds-manageable-list@0.12.1) (2020-06-05)
+
+**Note:** Version bump only for package @synerise/ds-manageable-list
+
+
+
+
+
 # [0.12.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-manageable-list@0.11.12...@synerise/ds-manageable-list@0.12.0) (2020-06-02)