+++ packages/components/modal/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.34](https://github.com/synerise/synerise-design/compare/@synerise/ds-modal@0.11.33...@synerise/ds-modal@0.11.34) (2020-06-05)
+
+**Note:** Version bump only for package @synerise/ds-modal
+
+
+
+
+
 ## [0.11.33](https://github.com/synerise/synerise-design/compare/@synerise/ds-modal@0.11.32...@synerise/ds-modal@0.11.33) (2020-06-03)