+++ packages/components/sidebar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.31](https://github.com/Synerise/synerise-design/compare/@synerise/ds-sidebar@0.2.30...@synerise/ds-sidebar@0.2.31) (2020-06-05)
+
+**Note:** Version bump only for package @synerise/ds-sidebar
+
+
+
+
+
 ## [0.2.30](https://github.com/Synerise/synerise-design/compare/@synerise/ds-sidebar@0.2.29...@synerise/ds-sidebar@0.2.30) (2020-06-02)
 
 **Note:** Version bump only for package @synerise/ds-sidebar